requirements.txt
pycryptodome>=3.10.4
requests>=2.26.0
python-dotenv >=0.18.0 
wheel
python-decouple>=3.4
python-dotenv>=0.19.0